User Story 7
As Player 1,
So I can lose a game of Battle,
I want Player 2 to attack me, and I want to get a confirmation

The Link on the confirmation page was also switched for a form containing a button, as clicking it now changes the `$game` state. It now uses POST to do this.